Add --host option to dev command (Closes #1) (#2)
When running the dev command on a Docker container, the default host
value of 127.0.0.1 that gets passed to the Webpack Dev Server will
allow connections only from within the container. In order to connect
from the docker host, the Webpack Dev Server needs to take 0.0.0.0 as
the host parameter.

This changeset introduces a new optional argument to the dev command,
`--host` that will take the address that will be passed as a host to
Webpack Dev Server. By default this 127.0.0.1.

Co-authored-by: Jacobo de Vera <jacobo.devera@hotjar.com>